requirements-test.txt
pytest<4
pytest-django
swh.scheduler[testing]
